refactor analyse/view (independent of #14868)
