+++ ui/analyse/src/view/main.ts
@@ -1,0 +1,79 @@
+import { view as cevalView } from 'ceval';
+import * as licon from 'common/licon';
+import { onInsert, looseH as h } from 'common/snabbdom';
+import { playable } from 'game';
+import * as router from 'game/router';
+import { VNode } from 'snabbdom';
+import { render as trainingView } from './roundTraining';
+import crazyView from '../crazy/crazyView';
+import AnalyseCtrl from '../ctrl';
+import forecastView from '../forecast/forecastView';
+import { view as keyboardView } from '../keyboard';
+import type * as studyDeps from '../study/studyDeps';
+import { relayView } from '../study/relay/relayView';
+import {
+  viewContext,
+  renderBoard,
+  renderMain,
+  renderControls,
+  renderTools,
+  renderUnderboard,
+} from './components';
+
+export default function (deps?: typeof studyDeps) {
+  return function (ctrl: AnalyseCtrl): VNode {
+    if (ctrl.nvui) return ctrl.nvui.render();
+    else if (ctrl.study?.relay) return relayView(ctrl, deps);
+    else return analyseView(ctrl, deps);
+  };
+}
+
+function analyseView(ctrl: AnalyseCtrl, deps?: typeof studyDeps): VNode {
+  const ctx = viewContext(ctrl, deps);
+  const { study, menuIsOpen, gamebookPlayView, gaugeOn } = ctx;
+
+  return renderMain(ctx, [
+    ctrl.keyboardHelp && keyboardView(ctrl),
+    study && deps?.studyView.overboard(study),
+    renderBoard(ctx),
+    gaugeOn && cevalView.renderGauge(ctrl),
+    !menuIsOpen && crazyView(ctrl, ctrl.topColor(), 'top'),
+    gamebookPlayView || renderTools(ctx),
+    !menuIsOpen && crazyView(ctrl, ctrl.bottomColor(), 'bottom'),
+    !gamebookPlayView && renderControls(ctrl),
+    renderUnderboard(ctx),
+    trainingView(ctrl),
+    ctrl.studyPractice
+      ? deps?.studyPracticeView.side(study!)
+      : h(
+          'aside.analyse__side',
+          {
+            hook: onInsert(elm => {
+              ctrl.opts.$side && ctrl.opts.$side.length && $(elm).replaceWith(ctrl.opts.$side);
+            }),
+          },
+          ctrl.studyPractice
+            ? [deps?.studyPracticeView.side(study!)]
+            : study
+            ? [deps?.studyView.side(study, true)]
+            : [
+                ctrl.forecast && forecastView(ctrl, ctrl.forecast),
+                !ctrl.synthetic &&
+                  playable(ctrl.data) &&
+                  h(
+                    'div.back-to-game',
+                    h(
+                      'a.button.button-empty.text',
+                      {
+                        attrs: {
+                          href: router.game(ctrl.data, ctrl.data.player.color),
+                          'data-icon': licon.Back,
+                        },
+                      },
+                      ctrl.trans.noarg('backToGame'),
+                    ),
+                  ),
+              ],
+        ),
+  ]);
+}
